Pull out SQL formatting logic into formatTableRowsToSQL and start a test
